requirements.txt
streamlit>=1.24
requests
pandas
Pillow
streamlit-folium
folium
sentence-transformers
transformers
torch
numpy
